refactor: convert and restructure type
