+++ src/utils/edit/getValue.ts
@@ -1,6 +1,10 @@
 import {isContentEditable} from './isContentEditable'
 
-export function getValue(element: Element): string | null {
+export function getValue(element: Element | null): string | null {
+  // istanbul ignore if
+  if (!element) {
+    return null
+  }
   if (isContentEditable(element)) {
     return element.textContent
   }
