+++ src/utils/misc/index.ts
@@ -2,3 +2,4 @@ export * from './eventWrapper'
 export * from './isInstanceOfElement'
 export * from './isLabelWithInternallyDisabledControl'
 export * from './isVisible'
+export * from './isDisabled'
